requirements.txt
matplotlib
scipy
pillow
logzero